requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==1.11.7
django-ajax-selects==1.6.1
django-datatables-view==1.14.0
djangorestframework==3.7.3
gevent==1.2.2
greenlet==0.4.12
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
mysqlclient==1.3.10
pytz==2017.3
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
